fix: fixed an issue with relative paths by adding a new script to read in the username parts; also better for future improvements and work
